Hash PGO profile files to properly invalidate them
This PR changes merged PGO profile naming so that they will be named with the (BLAKE-3) hash of their contents. This will make sure that the compilation session will be invalidated if the profiles change (before this wouldn't happen, since the profiles used the same path in each compilation session).

This PR also fixes the merging of already merged profiles, which happened before. Now only the `.profraw` files will be merged. This should make it faster to run `cargo pgo optimize` if the code nor the profiles have changed.